chore(app-degree-pages): : clean file strcture up
