+++ packages/app-degree-pages/__mocks__/components-core-mock.js
