+++ packages/app-degree-pages/__mocks__/data/degree-search-mock.js
