+++ packages/app-degree-pages/__mocks__/data/input.json
