+++ packages/app-degree-pages/examples/program-detail-page.html
@@ -32,7 +32,7 @@
     <script src="../dist/degreePages.production.js"></script>
 
     <!--  MOCK RFI form component -->
-    <!-- <script src="../mocks/RequestInfoForm/index.js"></script> -->
+    <!-- <script src="../__mocks__/RequestInfoForm/index.js"></script> -->
     <style>
       div#my-request-form-info-id {
         position: relative;
